Add additional test case
