+++ src/Bicep.Core/Emit/ParametersJsonWriter.cs
@@ -3,83 +3,104 @@
 
 using Bicep.Core.Intermediate;
 using Bicep.Core.Semantics;
+using Microsoft.WindowsAzure.ResourceStack.Common.Json;
 using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 
 namespace Bicep.Core.Emit;
 
 public class ParametersJsonWriter
 {
-    private readonly SemanticModel model;
+    private ExpressionBuilder ExpressionBuilder { get; }
 
-    public ParametersJsonWriter(SemanticModel model)
+    private EmitterContext Context => ExpressionBuilder.Context;
+
+    public ParametersJsonWriter(SemanticModel semanticModel)
     {
-        this.model = model;
+        ExpressionBuilder = new ExpressionBuilder(new EmitterContext(semanticModel));
     }
 
-    public void Write(JsonTextWriter jsonWriter)
+    public void Write(SourceAwareJsonTextWriter writer)
     {
+        var parametersJToken = GenerateParametersJToken(writer.TrackingJsonWriter);
+        parametersJToken.WriteTo(writer);
+    }
+
+    private JToken GenerateParametersJToken(PositionTrackingJsonTextWriter jsonWriter)
+    {
+        var emitter = new ExpressionEmitter(jsonWriter, this.Context);
+        
         jsonWriter.WriteStartObject();
-
-        jsonWriter.WritePropertyName("$schema");
-        jsonWriter.WriteValue("https://schema.management.azure.com/schemas/2019-04-01/deploymentParameters.json#");
-
-        jsonWriter.WritePropertyName("contentVersion");
-        jsonWriter.WriteValue("1.0.0.0");
-
-        jsonWriter.WritePropertyName("parameters");
-        jsonWriter.WriteStartObject();
-
-        foreach (var assignment in model.Root.ParameterAssignments)
+        emitter.EmitProperty("$schema", "https://schema.management.azure.com/schemas/2019-04-01/deploymentParameters.json#");
+        emitter.EmitProperty("contentVersion", "1.0.0.0");
+        emitter.EmitObjectProperty("parameters", () =>
         {
-            jsonWriter.WritePropertyName(assignment.Name);
-            jsonWriter.WriteStartObject();
-
-            var parameter = model.EmitLimitationInfo.ParameterAssignments[assignment];
-
-            if (parameter.KeyVaultReferenceExpression is { } keyVaultReference)
+            foreach (var assignment in this.Context.SemanticModel.Root.ParameterAssignments)
             {
-                WriteKeyVaultReference(jsonWriter, keyVaultReference);
-            }
-            else if (parameter.Value is { } value)
-            {
-                jsonWriter.WritePropertyName("value");
-                value.WriteTo(jsonWriter);
-            }
-            else
-            {
-                throw new InvalidOperationException($"The '{assignment.Name}' parameter assignment defined neither a concrete value nor a key vault reference");
-            }
+                emitter.EmitObjectProperty(assignment.Name, () => 
+                {
+                    var parameter = this.Context.SemanticModel.EmitLimitationInfo.ParameterAssignments[assignment];
 
-            jsonWriter.WriteEndObject();
+                    if (parameter.KeyVaultReferenceExpression is { } keyVaultReference)
+                    {
+                        WriteKeyVaultReference(emitter, keyVaultReference);
+                    }
+                    else if (parameter.Value is { } value)
+                    {
+                        emitter.EmitProperty("value", () => value.WriteTo(jsonWriter));
+                    }
+                    else if (parameter.Expression is { } expression)
+                    {
+                        emitter.EmitProperty("expression", expression);
+                    }
+                    else
+                    {
+                        throw new InvalidOperationException($"The '{assignment.Name}' parameter assignment defined neither a concrete value nor a key vault reference");
+                    }
+                });
+            }
+        });
+        
+        if (this.Context.SemanticModel.Root.ParameterAssignments.Length > 0 &&
+            this.Context.ExternalInputReferences.ExternalInputIndexMap.Count > 0)
+        {
+            emitter.EmitObjectProperty("externalInputs", () =>
+            {
+                foreach (var reference in this.Context.ExternalInputReferences.ExternalInputIndexMap)
+                {
+                    var functionCall = reference.Key;
+                    var index = reference.Value;
+
+                    var expression = (FunctionCallExpression)ExpressionBuilder.Convert(functionCall);
+                    emitter.EmitObjectProperty(index.ToString(), () =>
+                    {
+                        emitter.EmitProperty("kind", expression.Parameters[0]);
+                        emitter.EmitProperty("options", expression.Parameters[1]);
+                    });
+                }
+            });
         }
 
         jsonWriter.WriteEndObject();
-
-        jsonWriter.WriteEndObject();
+        var content = jsonWriter.ToString();
+        return content.FromJson<JToken>();
     }
 
-    private static void WriteKeyVaultReference(JsonWriter jsonWriter, ParameterKeyVaultReferenceExpression expression)
+    private static void WriteKeyVaultReference(ExpressionEmitter emitter, ParameterKeyVaultReferenceExpression keyVaultReference)
     {
-        jsonWriter.WritePropertyName("reference");
-        jsonWriter.WriteStartObject();
-
-        jsonWriter.WritePropertyName("keyVault");
-        jsonWriter.WriteStartObject();
-
-        jsonWriter.WritePropertyName("id");
-        jsonWriter.WriteValue(expression.KeyVaultId);
-
-        jsonWriter.WriteEndObject();
-
-        jsonWriter.WritePropertyName("secretName");
-        jsonWriter.WriteValue(expression.SecretName);
-
-        if (expression.SecretVersion is string secretVersion)
+        emitter.EmitObjectProperty("reference", () =>
         {
-            jsonWriter.WritePropertyName("secretVersion");
-            jsonWriter.WriteValue(secretVersion);
-        }
+            emitter.EmitObjectProperty("keyVault", () =>
+            {
+                emitter.EmitProperty("id", keyVaultReference.KeyVaultId);
+            });
 
-        jsonWriter.WriteEndObject();
+            emitter.EmitProperty("secretName", keyVaultReference.SecretName);
+
+            if (keyVaultReference.SecretVersion is { } secretVersion)
+            {
+                emitter.EmitProperty("secretVersion", secretVersion);
+            }
+        });
     }
 }
